Polish pass: new SFX, sprite generators, levelgen fix, bug bounty expansion
- Add 8 new SFX (weapon-specific fire + enemy hit/death variety)
- Add boss_sprites.py and enemy_sprites.py sprite generator tools
- Fix levelgen decorations overwriting collision with set_visual_tile()
- Expand bug bounty to 7 tiers with difficulty modifiers (swarm/iron)
- Add 3 particle types (smoke, electric, heal)
- Add ability buff timer getters for HUD display
- Add video_reset_effects() for clean state transitions
- Add enemy spawn grace timer and entity-slot reverse mapping
- Track enemy max_hp at spawn for HP bar scaling
- Boss palette dirty tracking to avoid redundant recalculation
- Tune drone/turret base stats

+++ include/game/projectile.h
@@ -22,7 +22,7 @@ typedef struct {
     u8  owner;         /* Entity pool index of owner */
     u8  oam_index;     /* OAM sprite index */
     u8  pad;
-    u16 hit_mask;      /* Bitmask of enemy slots already hit (piercing dedup) */
+    u32 hit_mask;      /* Bitmask: bits 0-23 = enemy slots, bit 31 = boss (piercing dedup) */
 } Projectile;
 
 /* Initialize the projectile pool. */
